backend /requirements.txt
Django==4.1.7
djangorestframework
channels
psycopg2